+++ cookbooks/aws-parallelcluster-shared/spec/spec_helper.rb
@@ -41,6 +41,7 @@ def for_all_oses
     %w(amazon 2023),
     %w(ubuntu 22.04),
     %w(ubuntu 24.04),
+    %w(almalinux 8),
     %w(redhat 8),
     %w(rocky 8),
     %w(redhat 9),